API권한 수정 및 에러메시지 수정

+++ src/main/java/com/mighty/ninda/service/CommentService.java
@@ -5,6 +5,7 @@ import com.mighty.ninda.domain.comment.Comment;
 import com.mighty.ninda.domain.comment.CommentRepository;
 import com.mighty.ninda.domain.post.Post;
 import com.mighty.ninda.domain.post.PostRepository;
+import com.mighty.ninda.domain.user.RegistrationId;
 import com.mighty.ninda.domain.user.Role;
 import com.mighty.ninda.domain.user.User;
 import com.mighty.ninda.domain.user.UserRepository;
@@ -32,6 +33,8 @@ public class CommentService {
 
     @Transactional
     public void saveComment(CurrentUser currentUser, SaveComment requestDto) {
+
+        checkAuth(currentUser);
 
         Long userId = currentUser.getId();
 
@@ -97,23 +100,29 @@ public class CommentService {
 
     @Transactional
     public void updateComment(CurrentUser currentUser, Long commentId, UpdateComment requestDto) {
+
+        checkAuth(currentUser);
+
+        Long userId = currentUser.getId();
         Comment comment = findById(commentId);
 
-//        if (!sessionUser.getId().equals(comment.getUser().getId()) && sessionUser.getRole() != Role.ADMIN) {
-//            throw new HandleAccessDenied("작성자만 변경할 수 있습니다.");
-//        }
+        if (!userId.equals(comment.getUser().getId()) && currentUser.getRole() != Role.ADMIN) {
+            throw new HandleAccessDenied("작성자만 변경할 수 있습니다.");
+        }
 
         comment.update(requestDto.getContext());
     }
 
     @Transactional
     public void deleteComment(CurrentUser currentUser, Long commentId) {
+
+        checkAuth(currentUser);
+
+        Long userId = currentUser.getId();
         Comment comment = findById(commentId);
-
-//        if (!sessionUser.getId().equals(comment.getUser().getId()) && sessionUser.getRole() != Role.ADMIN) {
-//            throw new HandleAccessDenied("작성자만 삭제할 수 있습니다.");
-//        }
-
+        if (!userId.equals(comment.getUser().getId()) && currentUser.getRole() != Role.ADMIN) {
+            throw new HandleAccessDenied("작성자만 삭제할 수 있습니다.");
+        }
         commentRepository.delete(comment);
     }
 
@@ -121,11 +130,7 @@ public class CommentService {
     public void reLikeUp(CurrentUser currentUser, Long commentId) {
         Comment comment = findById(commentId);
 
-        if (currentUser == null) {
-            throw new HandleAccessDenied("로그인이 필요합니다.");
-        } else if (currentUser.getRole() == Role.GUEST) {
-            throw new HandleAccessDenied("아직 인증이 완료되지 않았습니다.");
-        }
+        checkAuth(currentUser);
 
         Long userId = currentUser.getId();
 
@@ -143,11 +148,7 @@ public class CommentService {
     public void reHateUp(CurrentUser currentUser, Long commentId) {
         Comment comment = findById(commentId);
 
-        if (currentUser == null) {
-            throw new HandleAccessDenied("로그인이 필요합니다.");
-        } else if (currentUser.getRole() == Role.GUEST) {
-            throw new HandleAccessDenied("아직 인증이 완료되지 않았습니다.");
-        }
+        checkAuth(currentUser);
 
         Long userId = currentUser.getId();
 
@@ -158,6 +159,18 @@ public class CommentService {
         } else {
             comment.reHateUp();
             comment.updateHateList(_userId);
+        }
+    }
+
+    public void checkAuth(CurrentUser currentUser) {
+        if (currentUser == null) {
+            throw new HandleAccessDenied("로그인이 필요한 서비스에요.");
+        } else if (currentUser.getRole() == Role.GUEST) {
+            if (currentUser.getRegistrationId() == RegistrationId.NINDA) {
+                throw new HandleAccessDenied("메일인증을 완료해주세요.");
+            } else {
+                throw new HandleAccessDenied("닉네임을 설정해주세요.");
+            }
         }
     }
 }
